requirements.txt
# LexiQuery - Intelligent Legal Contract Analysis System
# Install with: pip install -r requirements.txt

pdfplumber==0.11.4
pillow==10.4.0
chromadb==0.5.0
groq==0.9.0
sentence-transformers==3.0.1
fastapi==0.111.0
uvicorn==0.30.1
python-multipart==0.0.9
streamlit==1.36.0
python-dotenv==1.0.1
langchain==0.2.6
langchain-community==0.2.6